requirements.txt
streamlit>=1.29.0
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
bcrypt>=4.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
Pillow>=10.0.0
streamlit-option-menu>=0.3.0
streamlit-lottie>=0.0.5
pytz>=2023.3
xlsxwriter>=3.1.0
psycopg2-binary>=2.9.0